requirements.txt
streamlit
torch
torchvision
pillow
opencv-python
numpy
